third_party/WebKit/LayoutTests/shadow-dom/nodetree-labels-node-list.html

<!DOCTYPE html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
    var host = document.createElement('div');
    var root = host.attachShadow({mode: 'open'});
    root.innerHTML = '<input id="i1"><label for="i1"></label>';
    assert_equals(root.firstChild.labels.length, 1);

    document.body.appendChild(host);
    assert_equals(root.firstChild.labels.length, 1);
}, 'LabelsNodeList should work in Shadow DOM.');
</script>
